+++ Runtime/Extensions/ServiceLocator/TimerService.cs
@@ -1,6 +1,8 @@
-﻿namespace Nonatomic.Timers.Extensions.ServiceLocator
+﻿#if SERVICE_LOCATOR
+
+namespace Nonatomic.Timers.Extensions.ServiceLocator
 {
-	public interface ITimerService : ITimer
+	public interface ITimerService : IBaseTimerService, ITimer
 	{
 		
 	}
@@ -10,3 +12,5 @@
 		
 	}
 }
+
+#endif